pyproject.toml
[tool.poetry]
name = "trading-network-electronics-drf"
version = "0.1.0"
description = ""
authors = ["Kettariec <kettariec@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.3.1"
pillow = "^10.3.0"
djangorestframework-simplejwt = "^5.3.1"
rest-framework-simplejwt = "^0.0.2"
django-filter = "^24.2"
djangorestframework-filter = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
